+++ src/main/java/frc/robot/RobotContainer.java
@@ -11,7 +11,7 @@ import edu.wpi.first.wpilibj.GenericHID;
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.commands.ArcadeDrive;
+import frc.robot.commands.driving.ArcadeDrive;
 import frc.robot.subsystems.Drivetrain;
 
 /**
